==> 01_read_in_raw_data.ipynb <==
##### Purpose: to process data files from lexisuni 
##### Author: Julia Cope
##### Creation Date: 4/20/23
##### Project: A2 NLP - capturing climate claims 
##### Inputs: 'C:/01_Pr/' 
##### Above contains a folder in which is the datasets from Nexis Uni Bulk 
##### Inputs: 
##### Inputs: 
##### Output: '03_Outputs/01_text_metadata.csv'
##### Output: 
##### Output: 

import pandas as pd
import numpy as np

import csv


# Import Module
import os

# Folder Path
#path = "C:\\01_Pr"
path = "C:\\2P"

# Change the directory
os.chdir(path)

## print working direction
os.getcwd()

#https://stackoverflow.com/questions/5186839/python-replace-with 

#def get_df():
def get_df(path,s_string):
    #initialize empty df 
    #metadata_df = pd.DataFrame()
    metadata_df = []
    #data = []
    text_data_array = []
    text_filename_array = []
    column_names = ['Filename', 'Publication', 'Section', 'Date', 'Title', 'Author',
       'LNID']
    #rootdir = './'
    #rootdir = ".\\"
    #rootdir_full = 'C:/Users/julia/OneDrive/Documents/Penn 2022-2023/nlp project/press releases 1/julia.cope@asc.upenn.edu-b1720-s2023-03-16/'
    i = 0 
    #starts_with_string = rootdir_full + 'term_press_release'
    #starts_with_string = path + 'term_press_release'
    starts_with_string = path + s_string

    
    for subdir, dirs, files in os.walk(path):
        for file in files:
            #print("In folder: " + subdir)
            if subdir.startswith(starts_with_string):
                if file.endswith('.csv'):
                    #print("In folder: " + subdir)
                    csv_path = os.path.join(subdir, file)
                    csv_path = str(csv_path).replace(os.path.sep, '/')
                    #csv_path = str(subdir) + '/' + str(file)
                    #print(str(csv_path))
                    #csv_path
                    #print(os.path.join(subdir, file))
                    with open(str(csv_path), newline='',encoding="utf8") as csvfile:
                        temp_data1 = list(csv.reader(csvfile))
   
                    metadata_df.extend(temp_data1[1:])
                    #temp_meta_df = pd.read_csv(str(csv_path))
                    #metadata_df = pd.concat([metadata_df, temp_meta_df])
                    #print(metadata_df)
                if file.endswith('.txt'):
                    i = i + 1 
                    ## open file 
                    ## read the data 
                    ## put it in the numpy array 
                    txt_path = str(subdir).replace(os.path.sep, '/') + '/' + str(file)
                    
                    #print(txt_path)
                    text_file = open(txt_path, "r",encoding="utf8")
                    #read whole file to a string
                    data = text_file.read()
                    #close file
                    text_file.close()
                    
                    ##append to array data 
                    text_data_array.append(data)
                    ##append to filename array 
                    text_filename_array.append(str(file))
                    
    


    #merge dataframes on Filename and filename
    #https://numpy.org/doc/stable/reference/generated/numpy.column_stack.html
    #text_arr = np.column_stack((text_filename_array,text_data_array))
    d = dict(zip(text_filename_array, text_data_array))

    #d = {'Filename': text_filename_array, 'text': text_data_array}
    text_filename_array=[]
    text_data_array = []
    
    
    ## get metadata as a df
    metadata_df = pd.DataFrame(metadata_df)
    metadata_df.columns = column_names
    ## merge dfs
    metadata_df["text"] = metadata_df["Filename"].map(d)
    
    
    #articles_df_output = metadata_df.merge(text_df, on = "Filename", how="outer")
    print("There are " + str(i) + " articles")
    
    return metadata_df




#path = 'C:/Users/julia/OneDrive/Documents/Penn 2022-2023/nlp project/press releases 1/julia.cope@asc.upenn.edu-b1720-s2023-03-16/'
# 30 seconds for 1300 articles
## 10 minutes for 30,000 articles
#path = 'C:/01_Pr/'
path = 'C:/2P/Che/'

s_string = 'term_press_release'
PR01_df =  get_df(path,s_string)

## a few minutes
path = 'C:/2P/Mar1/'
s_string = 'term_press_release'
PR03_df = get_df(path,s_string)

print(PR03_df)

path = 'C:/2P/Val/'
s_string = 'term_press_release'
PR04_df = get_df(path,s_string)

##save as a csv file 
outputdf =   PR04_df                
#outputdf = pd.concat([ PR03_df, PR04_df])
#outputdf = pd.concat([PR01_df, PR03_df_2017])



#outputdf

path = "C:\\Users\\julia\\OneDrive\\Desktop\\NLP"

# Change the directory
os.chdir(path)
outputdf.to_csv('03_Outputs/01_valero_metadata.csv',index=False)

#outputdf.to_csv('03_Outputs/01_text_metadata.csv',index=False)
