==> 02_clean_dataframe.ipynb <==
##### Purpose: to process text data into sentencec level with index. 
##### Author: Julia Cope
##### Creation Date: 4/20/23
##### Project: A2 NLP - capturing climate claims 
##### Inputs: 
##### Inputs: 03_Outputs/01_text_metadata.csv

##### Inputs: 03_Outputs/01_exxon_metadata.csv
##### Inputs: 03_Outputs/01_chevron_metadata.csv
##### Inputs: 03_Outputs/01_marathon_metadata.csv
##### Inputs: 03_Outputs/01_phillips_metadata.csv
##### Inputs: 03_Outputs/01_valero_metadata.csv

##### Output:  
##### Output: 02_textsentences.csv
##### Output: 

import pandas as pd
import numpy as np
import nltk
from nltk.tokenize import sent_tokenize
import random
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.decomposition import LatentDirichletAllocation as LDA
import itertools
random.seed(10)

### split into train, test, dev
def func_split_three_dfs(df):
    
    df['randNumCol'] = pd.Series(range(1,4)).sample(16190, replace=True,random_state = 10).array
    df = df.reset_index()
    
    ## df has 16190 rows
    traindf = df[df['randNumCol'] == 1]
    ## traindf has 5359 rows
    devdf = df[df['randNumCol'] == 2]
    ##dev df has 5396 rows
    testdf = df[df['randNumCol'] == 3]
    ## testdf has 5435 rows
    
    return traindf, devdf, testdf

def func_split_metadata_text(df_ninecols):
    text_df = df_ninecols[['index','Date','text','company']]
    ## keep index
    
    return text_df

#def func_clean_data(text_df):
    ## remove the date first words
    
    ## separate paragraphs
    ## separate sentences
    
#cleaned_dev_text = func_clean_data(dev_text_df)
#text = dev_text_df["text"][200]

## read in data 
#text_metadata_df = pd.read_csv('03_Outputs/01_text_metadata.csv')


exx_df = pd.read_csv('03_Outputs/01_exxon_metadata.csv')
che_df = pd.read_csv('03_Outputs/01_chevron_metadata.csv')
mar_df = pd.read_csv('03_Outputs/01_marathon_metadata.csv')
phi_df = pd.read_csv('03_Outputs/01_phillips_metadata.csv')
val_df = pd.read_csv('03_Outputs/01_valero_metadata.csv')



exx_df['company'] = 'Exxon'
che_df['company'] = 'Chevron'
mar_df['company'] = 'Marathon'
phi_df['company'] = 'Phillips'
val_df['company'] = 'Valero'



text_metadata_df = pd.concat([exx_df, che_df, mar_df, phi_df, val_df])

text_metadata_df


text_metadata_df['text'] = text_metadata_df['text'].astype(str)
#traindf, devdf, testdf = func_split_three_dfs(text_metadata_df)
#dev_text_df = func_split_metadata_text(devdf)
text_metadata_df = text_metadata_df.reset_index()
full_text_df = func_split_metadata_text(text_metadata_df)

full_text_df

# Function to split text into sentences
def split_sentences(text):
    return nltk.sent_tokenize(text)

# Apply the function to the text column
#dev_text_df['sentences'] = dev_text_df['text'].apply(split_sentences)
full_text_df['sentences'] = full_text_df['text'].apply(lambda x: nltk.sent_tokenize(x))




del full_text_df['text']

## takes about 4 to 5 minutes
full_text_df = full_text_df.explode('sentences')



full_text_df

#text_metadata_df

### export file

import os

path = "C:\\Users\\julia\\OneDrive\\Desktop\\NLP"

# Change the directory
os.chdir(path)
full_text_df.to_csv('03_Outputs/02_PR_sentences.csv',index=False)
#full_text_df.to_csv('03_Outputs/02_textsentences.csv',index=False)

### Below is experimentation...
