==> 03_Explore_ClimateFever.ipynb <==
##### Purpose:  
##### Author: Julia Cope
##### Creation Date: 4/20/23
##### Project: A2 NLP - capturing climate claims 
##### Inputs: 
##### Inputs: 
##### Inputs: 
##### Output: 
##### Output: 
##### Output: 

import os
import pandas as pd
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.decomposition import LatentDirichletAllocation as LDA


pd.options.display.max_colwidth = 150

def func_make_evidence_df(df, ID_column, column_name):
    
    evidence_df = df[[ID_column, column_name]].explode(column_name)
    temp_df = pd.json_normalize(evidence_df[column_name])
    temp_df.reset_index(inplace=True, drop=True)
    evidence_df.reset_index(inplace=True, drop=True)
    evidence_df = pd.concat([evidence_df['claim_id'], temp_df], axis=1)
    return evidence_df

## https://scikit-learn.org/stable/auto_examples/applications/plot_topics_extraction_with_nmf_lda.html#sphx-glr-auto-examples-applications-plot-topics-extraction-with-nmf-lda-py
def print_top_words(H, vocab, n_top_words, sample_topics):
    for topic_idx, topic in enumerate(H): ##(iterate thru each topic)
        ##this gives the indexes of the highest values in each topic word row 
        ##(which words are in the topic most often)
        top_features_ind = topic.argsort()[: -n_top_words - 1 : -1]
        ##in the vocab (aka feature_names) this is the actual list of the string of words
        top_features = [vocab[i] for i in top_features_ind]
        ## weights are whatever the value in the H matrix are - im not super sure of the meaning
        weights = topic[top_features_ind]
        ##gets each word  in the topic from matrix H
        if topic_idx in sample_topics:
            print(str(topic_idx) + str(top_features))
            print("\n")
        ##actually save the topics instead of printing them- and give them little numbers
        
        
def topics_per_doc(W, n_top_topics, samplearts, H, vocab, n_top_words, titles):
    for topic_idx, topic in enumerate(W): ##(iterate thru each document)
        ##this gives the indexes of the highest values in each topic word row 
        ##(which words are in the topic most often)
        top_features_ind = topic.argsort()[: -n_top_topics - 1 : -1]
        ##in the vocab (aka feature_names) this is the actual list of the string of words
        top_features = [i for i in top_features_ind]
        ## replace feature_names with the topics tnames that Im gonna make up
        ##top_features = [feature_names[i] for i in top_features_ind]
        ## weights are whatever the value in the H matrix are - im not super sure of the meaning
        weights = topic[top_features_ind]
        ##gets each word  in the topic from matrix H
        if topic_idx in samplearts:
            print("article: " + titles[topic_idx])
            ##get the article title from the titles list
            ## get the words from the words per topic function
            print_top_words(H, vocab, n_top_words, top_features)
            print("\n")


os.getcwd()
filename = 'C:/Users/julia/OneDrive/Desktop/NLP/climate-fever.jsonl'
climate_claims_df_INPUT = pd.read_json(filename, lines=True)
climate_claims_df = climate_claims_df_INPUT
## 1535 rows
climate_claims_df.head()

climate_claims_df.astype({'claim': 'string', 'claim_label':'category'}).dtypes
climate_claims_df.info()

evidence_df = func_make_evidence_df(climate_claims_df, 'claim_id', 'evidences')
climate_claims_df = climate_claims_df.drop(columns=['evidences'])

evidence_df

climate_claims_df[climate_claims_df['claim_label'] == "SUPPORTS"]

evidence_df[evidence_df['claim_id'] == 3134]

evidence_df[evidence_df["entropy"] > 0.7]

evidence_df['evidence']


climate_claims_df['claim']

#climate_claims_df['claim'] = climate_claims_df['claim'].str.replace('carbon dioxide', 'carbon_dioxide')
#climate_claims_df['claim'] = climate_claims_df['claim'].str.replace('climate change', 'climate_change')
#climate_claims_df['claim'] = climate_claims_df['claim'].str.replace('global warming', 'global_warming')

vectorizer = CountVectorizer(min_df=0.05, max_df=0.7,ngram_range = (1,5), strip_accents = 'unicode',stop_words='english')
dtm = vectorizer.fit_transform(evidence_df['evidence'])
vocab = vectorizer.get_feature_names()
used_stop_words = vectorizer.get_stop_words()
## 35 topics
ldamodel = LDA(n_components=10)
## 
W = ldamodel.fit_transform(dtm) #(documents x topics)
H = ldamodel.components_ #(topics x words)
print(vocab)

#print(used_stop_words)

#PRsample = [1,2,3,4]
#PRsampID = climate_claims_df[climate_claims_df['claim_id'].isin(PRsample)].index.tolist()
#topics_per_doc(W, 3, PRsampID, H, vocab, 10, climate_claims_df['claim'].tolist())


print_top_words(H, vocab, 3, list(range(0, H.shape[0])))  



## sklearn DTM creation - document-term matrix
## info from https://stackoverflow.com/questions/39931941/create-a-dtm-from-large-corpus
## TF IDF is Term Frequency(TF) and Inverse Document Frequency(IDF)

## min df - a given word has to be in some % of articles (try 0.05)(originally I did 0.12)
## max df - Baird has set it lower than 0.90 in the past - couldbe 0.60 - 
## try printng out words that have doc freq of > 0.6 - are they informative? 
## start with n grams of 1 - and then add complexity 
## 
##vectorizer = CountVectorizer(min_df=0.05,max_df=0.95, ngram_range = (1,1), strip_accents = 'unicode', stop_words='english')
vectorizer = CountVectorizer(min_df=0.05, max_df=0.7,ngram_range = (1,5), strip_accents = 'unicode',stop_words='english')
dtm = vectorizer.fit_transform(climate_claims_df['claim'])
vocab = vectorizer.get_feature_names_out()
used_stop_words = vectorizer.get_stop_words()

#print(used_stop_words)

# topic model
## 35 topics
ldamodel = LDA(n_components=10)
## 
W = ldamodel.fit_transform(dtm) #(documents x topics)
H = ldamodel.components_ #(topics x words)


print(len(vocab))
#vocab

print(vocab)

PRsample = [1,2,3,4]
PRsampID = climate_claims_df[climate_claims_df['claim_id'].isin(PRsample)].index.tolist()
topics_per_doc(W, 3, PRsampID, H, vocab, 10, climate_claims_df['claim'].tolist())


print_top_words(H, vocab, 3, list(range(0, H.shape[0])))
