==> 04_select sentences for labels.ipynb <==
# Beginning
+ header
+ libraries
+ read in data 



##### Purpose:  to take in setences from press releases and select a stratified random sample for hand coding.
##### Author: Julia Cope
##### Creation Date: 4/20/23
##### Project: A2 NLP - capturing climate claims 
##### Inputs: 
##### Inputs: 02_textsentences.csv
##### Inputs: 02_PR_sentences.csv


##### Output: 04_traindata
##### Output: 04_sampledata
##### Output: 

### libraries
import os
import pandas as pd
from sklearn.feature_extraction.text import CountVectorizer
import numpy as np
from sklearn.model_selection import train_test_split
import random


### read in data 
#sent_df = pd.read_csv('03_Outputs/02_textsentences.csv',parse_dates=['Date'])

#sent_df = pd.read_csv('03_Outputs/02_PR_sentences.csv',parse_dates=['Date'])

sent_df = pd.read_csv('03_Outputs/04B_unlabeled_data.csv',parse_dates=['Date'])


sent_df['Year'] = sent_df['Date'].dt.year 


sent_df = sent_df.dropna(subset=['sentences'])

sent_df['sentences'] = sent_df['sentences'].astype(str)

sent_df

# Get climate score 

+ save vocab
+ use count vectorizer to get score from vocab list 
+ score the sentences



vocab_climate_long = ['atmosphere', 'carbon', 'carbon dioxide',
                      'climate', 'climate change', 'co2','dioxide', 'earth', 'emissions', 
                      'global warming', 'greenhouse', 'greenhouse gases', 'human', 'ice', 
                      'level', 'ocean', 'rise', 'scientists', 'sea', 'sea level', 'surface', 
                      'temperature', 'temperatures', 'warming', 
                      
                      'aerosol' , "alternative energy" , 'biofuel' , "black carbon" , 
                      'carbon capture', 'carbon footprint' , 'carbon sequestration' ,
                     'ch4' ,  'climate feedback' , 'climate lag' , 'climate science' , 
                      'climate sensitivity' , 'coal' , 
                      "coal bleaching" , 'deforestation' , 'desertification' , 'ecology' , 'ecosystem' , 'emission' , 
                      "energy efficiency" , "enteric fermentation" , "environment" , "environmental", "extreme weather" , 
                      "fluorinated gas" , 'fluorocarbon' , "fossil fuel" , "fuel switching" , 'geothermal' , 'ghg' , 
                      'global temperature', 'greenhouse gas', "environmentally",
                      'greenhouse effect', 'halocarbon', 'hcfcs', 
                      "heat dome" , 'hfcs' , 'hydrochlorofluorocarbon' , 'hydrofluorocarbon' , 'hydropower' , "ice loss", 
                      'ipcc' , 'landfill' , 'methane' , "municipal solid waste" , 'n2o' , "nitrogen oxide" , 
                      "nitrous oxide" , 'nox' , "ocean acid" , 'ozone' , 
                      "particulate matter" , 'perfluorocarbon' , 'pfcs' , 
                      'pollution' , 'recycle' , 'reforest' , "resource management" , 'sf6' , 'solar' , 
                      "sulfur hexafluoride" , 'sustainability',
                      'sustainable' , 'renewable', "trace gas" , 'unfccc']

sent_df

# Split and Export


sent_df.head()

# get max length of values in 'text_col'
max_len = max(sent_df['sentences'].map(len))

print(max_len)

### select 5 sentences per year 

#random.seed(42)
train, test = train_test_split(sent_df, train_size=600, stratify=sent_df[['Year','company']], random_state=7)



# Print the shape of the training and testing sets
print('Training set shape:', train.shape)
print('Testing set shape:', test.shape)

### export the sample for hand coding 

## this count vectorizer checks how many words in the text are in my crafted vocab... 
## approx one minute

vectorizer = CountVectorizer(min_df=0.10, max_df=0.8,ngram_range = (1,3), vocabulary = vocab_climate_long,
                             strip_accents = 'unicode',stop_words='english')
dtm = vectorizer.fit_transform(train['sentences'])

## sum each row of the DTM to get a climate score  
array_score = np.sum(dtm, axis=1)
train['climate_score'] = array_score.tolist()
train = train.explode('climate_score')

train['likely'] = train['climate_score'].apply(lambda x: 1 if x > 1 else 0)

train

### export file
path = "C:\\Users\\julia\\OneDrive\\Desktop\\NLP"

# Change the directory
os.chdir(path)
train.to_csv('03_Outputs/04C_labeled_sampledata.csv',index=False)
test.to_csv('03_Outputs/04C_unlabeled_data.csv',index=False)

#train[train['climate_score'] > 0]
