==> 04B_combine files.ipynb <==
##### Purpose:  combine 04 labeled sentences into one file. 
##### Author: Julia Cope
##### Creation Date: 4/30/23
##### Project: A2 NLP - capturing climate claims 
##### Inputs: 04_labeled_sampledata.csv
##### Inputs: 04B_labeled_sampledata.csv
##### Inputs: 04C_labeled_sampledata.csv
##### Output: 
##### Output: 
##### Output:

### libraries
import os
import pandas as pd

### read in data 
df_A = pd.read_csv('03_Outputs/04_labeled_sampledata.csv',parse_dates=['Date'])
#sent_df = pd.read_csv('03_Outputs/02_PR_sentences.csv',parse_dates=['Date'])
df_A['Year'] = df_A['Date'].dt.year 

df_B = pd.read_csv('03_Outputs/04B_labeled_sampledata.csv',parse_dates=['Date'])
#sent_df = pd.read_csv('03_Outputs/02_PR_sentences.csv',parse_dates=['Date'])
df_B['Year'] = df_B['Date'].dt.year 

df_C = pd.read_csv('03_Outputs/04C_labeled_sampledata.csv',parse_dates=['Date'])
#sent_df = pd.read_csv('03_Outputs/02_PR_sentences.csv',parse_dates=['Date'])
df_C['Year'] = df_C['Date'].dt.year 

df_A.columns.tolist()

df_B.columns.tolist()

df_C.columns.tolist()

df_concat = pd.concat([df_A, df_B])

df_concat = pd.concat([df_concat, df_C])

### export file
path = "C:\\Users\\julia\\OneDrive\\Desktop\\NLP"

# Change the directory
os.chdir(path)
df_concat.to_csv('03_Outputs/04D_FULL_labeled_sampledata.csv',index=False)
