==> 05_featurize sentences .ipynb <==
# Beginning
+ header
+ libraries
+ read in data 



##### Purpose:  featurize the sentences, and make a naive bayes classifier
##### Author: Julia Cope
##### Creation Date: 4/20/23
##### Project: A2 NLP - capturing climate claims 
##### Inputs: 
##### Inputs: 04_sampledata.csv
##### Inputs: 04_labeled_sampledata.csv
##### Output: 
##### Output: 
##### Output: 

#!pip install spacy
#!python -m spacy download en_core_web_sm

### libraries
import os
import pandas as pd
#from sklearn.feature_extraction.text import CountVectorizer
import numpy as np
#from sklearn.model_selection import train_test_split
#import tensorflow_hub as hub
import spacy

### read in data 
sent_df = pd.read_csv('03_Outputs/04D_FULL_labeled_sampledata.csv',parse_dates=['Date'])
#sent_df = pd.read_csv('03_Outputs/04_labeled_sampledata.csv',parse_dates=['Date'])
sent_df['Year'] = sent_df['Date'].dt.year 


sent_df.head()

# Load pre-trained model
#30 seconds

nlp = spacy.load("en_core_web_sm")


# Define function to obtain sentence embeddings
def get_sentence_embedding(sentence):
    doc = nlp(sentence)
    return doc.vector

# Obtain sentence embeddings and create new dataframe with embeddings
embeddings = sent_df['sentences'].apply(get_sentence_embedding).apply(pd.Series)
df_embeddings = pd.concat([sent_df, embeddings], axis=1)

df_embeddings

### export file
path = "C:\\Users\\julia\\OneDrive\\Desktop\\NLP"

# Change the directory
os.chdir(path)
#df_embeddings.to_csv('03_Outputs/05_df_embeddings.csv',index=False)
df_embeddings.to_csv('03_Outputs/05_PR_embeddings.csv',index=False)
